Moved Login and LoginErrors to their own directory.
