+++ frontend/jobapptrackerfrontend/src/model/interfaces/login/Login.ts
