+++ frontend/jobapptrackerfrontend/src/model/interfaces/login/LoginErrors.ts
